String interning on by default
Also revert the Options changes. If the app needs to customize string interning, they can do so by changing the resolver to their own formatter.
