+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Formatters/StringInterningFormatter.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) All contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+using Microsoft.NET.StringTools;
+
 namespace MessagePack.Formatters
 {
     /// <summary>
@@ -8,6 +11,12 @@ namespace MessagePack.Formatters
     /// </summary>
     public sealed class StringInterningFormatter : IMessagePackFormatter<string>
     {
+        public static readonly StringInterningFormatter Instance = new StringInterningFormatter();
+
+        private StringInterningFormatter()
+        {
+        }
+
         /// <inheritdoc/>
         public string Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
@@ -16,12 +25,35 @@ namespace MessagePack.Formatters
                 return null;
             }
 
-            if (options.StringInterning is null)
+            MessagePackReader retryReader = reader;
+            if (reader.TryReadStringSpan(out ReadOnlySpan<byte> bytes))
             {
-                return reader.ReadString();
+                if (bytes.Length < 4096)
+                {
+                    Span<char> chars = stackalloc char[bytes.Length];
+                    int charLength;
+#if SPAN_BUILTIN
+                    charLength = StringEncoding.UTF8.GetChars(bytes, chars);
+#else
+                    unsafe
+                    {
+                        fixed (byte* pBytes = bytes)
+                        fixed (char* pChars = chars)
+                        {
+                            charLength = StringEncoding.UTF8.GetChars(pBytes, bytes.Length, pChars, chars.Length);
+                        }
+                    }
+#endif
+                    return Strings.WeakIntern(chars.Slice(0, charLength));
+                }
+                else
+                {
+                    // Rewind the reader to the start of the string because we're taking the slow path.
+                    reader = retryReader;
+                }
             }
 
-            return options.StringInterning.GetString(ref reader);
+            return Strings.WeakIntern(reader.ReadString());
         }
 
         /// <inheritdoc/>
